UX-237 Move Panel to legacy folder
