+++ packages/matchbox/src/components/Panel/legacy/Accent.js
