+++ packages/matchbox/src/components/Panel/legacy/Section.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { buttonsFrom } from '../Button';
+import { buttonsFrom } from '../../Button';
 import styled from 'styled-components';
 import { body, sectionContent, actions } from './styles';
 import { padding } from 'styled-system';
