+++ packages/matchbox/src/components/Panel/legacy/styles.js
